# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * JNX Production Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type GarmentType = typeof GarmentType[keyof typeof GarmentType];


export const GarmentType = {
  oversized_hoodie: 'oversized_hoodie',
  wide_cargo: 'wide_cargo',
  boxy_tee: 'boxy_tee',
  track_jacket: 'track_jacket',
  tailored_short: 'tailored_short',
} as const;

export type ValidationState = typeof ValidationState[keyof typeof ValidationState];


export const ValidationState = {
  PRELIMINARY_UNVALIDATED: 'PRELIMINARY_UNVALIDATED',
  SAMPLE_IN_REVIEW: 'SAMPLE_IN_REVIEW',
  APPROVED: 'APPROVED',
} as const;

export interface Color {
  name: string;
  hex: string;
  pantone: string;
}

export interface Measurement {
  code: string;
  name: string;
  valueCm: number;
  toleranceCm: number;
  gradeRule: string;
}

export interface Material {
  name: string;
  specification: string;
  supplier: string;
}

export interface Artwork {
  name: string;
  placement: string;
  notes: string;
}

export interface Template {
  id: GarmentType;
  name: string;
  description: string;
  measurements: Measurement[];
}

export interface ProjectInput {
  /** @minLength 1 */
  styleName: string;
  /** @minLength 1 */
  styleNumber: string;
  garmentType: GarmentType;
  baseSize: string;
  season: string;
  fit: string;
  description: string;
  concept: string;
  primaryColor: Color;
  measurements: Measurement[];
  materials: Material[];
  artworks: Artwork[];
}

export type Project = ProjectInput & {
  id: string;
  revision: number;
  validationState: ValidationState;
  createdAt: string;
  updatedAt: string;
};

export type PreviewAssetKind = typeof PreviewAssetKind[keyof typeof PreviewAssetKind];


export const PreviewAssetKind = {
  flat: 'flat',
  pattern: 'pattern',
} as const;

export interface PreviewAsset {
  kind: PreviewAssetKind;
  svg: string;
  warning: string;
}

export type GradingRowSizes = {[key: string]: number};

export interface GradingRow {
  code: string;
  name: string;
  sizes: GradingRowSizes;
  toleranceCm: number;
}

export interface FactoryPackage {
  packageName: string;
  downloadUrl: string;
  files: string[];
  validationState: ValidationState;
  warning: string;
}

export type ProjectSummaryGarmentCounts = {[key: string]: number};

export interface ProjectSummary {
  totalProjects: number;
  preliminaryProjects: number;
  garmentCounts: ProjectSummaryGarmentCounts;
  latestProject: Project | null;
}

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;